gcp-terraform/app/yolo/requirements.txt
dill
torch
ultralytics==8.1.23
opencv-python-headless==4.7.0.72
requests
flask
google-cloud-storage